pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openresponses-python"
version = "0.1.0"
description = "Official Python SDK for the Open Responses standard"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Uday", email = "uday@example.com" }
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]
openai = [
    "openai>=1.0.0",
]
all = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0", 
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "hatch>=1.7.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/uday/openresponses-python"

[tool.hatch.build.targets.wheel]
packages = ["src/openresponses"]
